Воссоздал проект WebStore.DAL, добавил все зависимости, перенес DbContext, перепсал его по другому
