+++ WebStore.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore", "WebStore\WebSto
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore.Domain", "WebStore.Domain\WebStore.Domain.csproj", "{F4CF666F-60AF-4956-A737-A1A42C7C5C19}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebStore.DAL", "WebStore.DAL\WebStore.DAL.csproj", "{890C10A5-D7AB-4503-9917-19E2B1236C1A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{F4CF666F-60AF-4956-A737-A1A42C7C5C19}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F4CF666F-60AF-4956-A737-A1A42C7C5C19}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F4CF666F-60AF-4956-A737-A1A42C7C5C19}.Release|Any CPU.Build.0 = Release|Any CPU
+		{890C10A5-D7AB-4503-9917-19E2B1236C1A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{890C10A5-D7AB-4503-9917-19E2B1236C1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{890C10A5-D7AB-4503-9917-19E2B1236C1A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{890C10A5-D7AB-4503-9917-19E2B1236C1A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
